some stuff I spotted in routes
These don't fix the whole problem, but they should get you going in the right direction, I think.

So, on line 13, your function was named `cate`, but you're exporting a function named `category` on line 6.

On line 16, `cate` was in quotes, making it a string. I imagine you wanted it as a parameter for the callback to `then`?